refactor: Move ComponentAuditTable
